Add files via upload
This is a full documentation of the AVL Tree List data structure (in Hebrew).
It has full analysis of methods and explanation about them and the time complexity of each method. 
Also it provides a comparison between this way of implementing a list and LL and arrays, and shows advantages and disadvantages for each data structure.